# divination-service/src/main/kotlin/com/github/butvinmitmo/divinationservice/controller/SpreadController.kt
package com.github.butvinmitmo.divinationservice.controller

import com.github.butvinmitmo.divinationservice.service.DivinationService
import com.github.butvinmitmo.shared.dto.CreateSpreadRequest
import com.github.butvinmitmo.shared.dto.CreateSpreadResponse
import com.github.butvinmitmo.shared.dto.SpreadDto
import com.github.butvinmitmo.shared.dto.SpreadSummaryDto
import io.swagger.v3.oas.annotations.Operation
import io.swagger.v3.oas.annotations.Parameter
import io.swagger.v3.oas.annotations.headers.Header
import io.swagger.v3.oas.annotations.media.ArraySchema
import io.swagger.v3.oas.annotations.media.Content
import io.swagger.v3.oas.annotations.media.Schema
import io.swagger.v3.oas.annotations.responses.ApiResponse
import io.swagger.v3.oas.annotations.responses.ApiResponses
import io.swagger.v3.oas.annotations.tags.Tag
import jakarta.validation.Valid
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import org.springframework.http.HttpStatus
import org.springframework.http.ResponseEntity
import org.springframework.validation.annotation.Validated
import org.springframework.web.bind.annotation.DeleteMapping
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestHeader
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.RestController
import reactor.core.publisher.Mono
import java.util.UUID

@RestController
@RequestMapping("/api/v0.0.1/spreads")
@Tag(name = "Spreads", description = "Tarot spread management and viewing")
@Validated
class SpreadController(
    private val divinationService: DivinationService,
) {
    @PostMapping
    @Operation(
        summary = "Create a new spread",
        description =
            "Creates a new tarot spread with randomly selected cards. Cards may be reversed. " +
                "ID and timestamp are generated by PostgreSQL. Returns only the generated ID.",
    )
    @ApiResponses(
        value = [
            ApiResponse(
                responseCode = "201",
                description = "Spread created successfully, returns generated ID",
                content = [Content(schema = Schema(implementation = CreateSpreadResponse::class))],
            ),
            ApiResponse(
                responseCode = "404",
                description = "Author or layout type not found",
                content = [
                    Content(
                        schema = Schema(implementation = com.github.butvinmitmo.shared.dto.ErrorResponse::class),
                    ),
                ],
            ),
            ApiResponse(
                responseCode = "400",
                description = "Invalid request data",
                content = [
                    Content(
                        schema =
                            Schema(
                                implementation = com.github.butvinmitmo.shared.dto.ValidationErrorResponse::class,
                            ),
                    ),
                ],
            ),
            ApiResponse(
                responseCode = "401",
                description = "Missing or invalid authentication",
                content = [
                    Content(
                        schema = Schema(implementation = com.github.butvinmitmo.shared.dto.ErrorResponse::class),
                    ),
                ],
            ),
        ],
    )
    fun createSpread(
        @Parameter(description = "User ID from JWT", required = true)
        @RequestHeader("X-User-Id")
        userId: UUID,
        @Parameter(description = "User role from JWT", required = true)
        @RequestHeader("X-User-Role")
        role: String,
        @Valid @RequestBody request: CreateSpreadRequest,
    ): Mono<ResponseEntity<CreateSpreadResponse>> =
        divinationService
            .createSpread(request, userId, role)
            .map { response -> ResponseEntity.status(HttpStatus.CREATED).body(response) }

    @GetMapping
    @Operation(
        summary = "Get paginated spreads",
        description = "Retrieves spreads sorted by creation date. Returns total count in X-Total-Count header.",
    )
    @ApiResponses(
        value = [
            ApiResponse(
                responseCode = "200",
                description = "Spreads retrieved successfully",
                headers = [
                    Header(
                        name = "X-Total-Count",
                        description = "Total number of spreads",
                        schema = Schema(type = "integer", implementation = Int::class),
                    ),
                ],
                content = [Content(array = ArraySchema(schema = Schema(implementation = SpreadSummaryDto::class)))],
            ),
        ],
    )
    fun getSpreads(
        @Parameter(description = "Page number (0-based)", example = "0")
        @RequestParam(defaultValue = "0")
        @Min(0)
        page: Int,
        @Parameter(description = "Page size (max 50)", example = "20")
        @RequestParam(defaultValue = "20")
        @Min(1)
        @Max(50)
        size: Int,
    ): Mono<ResponseEntity<List<SpreadSummaryDto>>> =
        divinationService
            .getSpreads(page, size)
            .map { response ->
                ResponseEntity
                    .ok()
                    .header("X-Total-Count", response.totalElements.toString())
                    .body(response.content)
            }

    @GetMapping("/scroll")
    @Operation(
        summary = "Get spreads with infinite scroll",
        description =
            "Retrieves spreads using cursor-based pagination. " +
                "Returns X-After header with cursor for next page if more items exist.",
    )
    @ApiResponses(
        value = [
            ApiResponse(
                responseCode = "200",
                description = "Spreads retrieved successfully",
                headers = [
                    Header(
                        name = "X-After",
                        description = "Cursor for next page (only present if more items exist)",
                        schema = Schema(type = "string", format = "uuid", implementation = String::class),
                    ),
                ],
                content = [Content(array = ArraySchema(schema = Schema(implementation = SpreadSummaryDto::class)))],
            ),
        ],
    )
    fun getSpreadsByScroll(
        @Parameter(description = "ID of the last spread from previous request (cursor)", required = false)
        @RequestParam(required = false)
        after: UUID?,
        @Parameter(description = "Number of spreads to retrieve (max 50)", example = "20")
        @RequestParam(defaultValue = "20")
        @Min(1)
        @Max(50)
        size: Int,
    ): Mono<ResponseEntity<List<SpreadSummaryDto>>> =
        divinationService
            .getSpreadsByScroll(after, size)
            .map { result ->
                val response = ResponseEntity.ok()
                if (result.nextCursor != null) {
                    response.header("X-After", result.nextCursor.toString()).body(result.items)
                } else {
                    response.body(result.items)
                }
            }

    @GetMapping("/{id}")
    @Operation(
        summary = "Get spread details",
        description = "Retrieves complete details of a specific spread including all cards and interpretations",
    )
    @ApiResponses(
        value = [
            ApiResponse(
                responseCode = "200",
                description = "Spread found",
                content = [Content(schema = Schema(implementation = SpreadDto::class))],
            ),
            ApiResponse(
                responseCode = "404",
                description = "Spread not found",
                content = [
                    Content(
                        schema = Schema(implementation = com.github.butvinmitmo.shared.dto.ErrorResponse::class),
                    ),
                ],
            ),
        ],
    )
    fun getSpread(
        @Parameter(description = "Spread ID", required = true)
        @PathVariable
        id: UUID,
    ): Mono<ResponseEntity<SpreadDto>> =
        divinationService
            .getSpread(id)
            .map { spread -> ResponseEntity.ok(spread) }

    @DeleteMapping("/{id}")
    @Operation(
        summary = "Delete a spread",
        description = "Deletes a spread and all associated data. Only the author can delete it.",
    )
    @ApiResponses(
        value = [
            ApiResponse(responseCode = "204", description = "Spread deleted successfully"),
            ApiResponse(
                responseCode = "404",
                description = "Spread not found",
                content = [
                    Content(
                        schema = Schema(implementation = com.github.butvinmitmo.shared.dto.ErrorResponse::class),
                    ),
                ],
            ),
            ApiResponse(
                responseCode = "403",
                description = "User is not the author of the spread",
                content = [
                    Content(
                        schema = Schema(implementation = com.github.butvinmitmo.shared.dto.ErrorResponse::class),
                    ),
                ],
            ),
            ApiResponse(
                responseCode = "401",
                description = "Missing or invalid authentication",
                content = [
                    Content(
                        schema = Schema(implementation = com.github.butvinmitmo.shared.dto.ErrorResponse::class),
                    ),
                ],
            ),
        ],
    )
    fun deleteSpread(
        @Parameter(description = "Spread ID", required = true)
        @PathVariable
        id: UUID,
        @Parameter(description = "User ID from JWT", required = true)
        @RequestHeader("X-User-Id")
        userId: UUID,
        @Parameter(description = "User role from JWT", required = true)
        @RequestHeader("X-User-Role")
        role: String,
    ): Mono<ResponseEntity<Void>> =
        divinationService
            .deleteSpread(id, userId, role)
            .then(Mono.just(ResponseEntity.noContent().build()))
}